# feature: async digest handling