--- a/src/dune_engine/import.ml
+++ b/src/dune_engine/import.ml
@@ -9,6 +9,7 @@ module Persistent = Dune_util.Persistent
 module Execution_env = Dune_util.Execution_env
 module Glob = Dune_glob.V1
 include No_io
+include Dune_config
 
 (* To make bug reports usable *)
 let () = Printexc.record_backtrace true