Move the computation of digests to a background thread. For now, this is guarded behind a configuration option that requires this to be enabled.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: fa4d841b-d0ce-412e-b84b-d2d53c0d7eb6 -->