@Alizter 2023-03-21
Then bench shows a noticeable slowdown:

Compare main:
```
1 | 0.9362449645996094
2 | 0.9103691577911377
3 | 0.9048709869384766
4 | 0.9169681072235107
5 | 0.925853967666626
```
with this pr:
```
1 | 0.9660398960113525
2 | 0.9446120262145996
3 | 0.9524080753326416
4 | 0.9530608654022217
5 | 0.9539759159088135
```

Also the clear build times make this clearer:
```
Main: 107.39912676811218
PR:   111.15323400497437
```