@rgrinberg 2023-03-21
Thanks for the initial numbers. This PR is waiting for a change in the way we handle digests to be more multi threading friendly.