@Alizter 2023-04-18
#### New bench results:
| | main | PR | %diff |
| --- | --- | --- | --- |
| Clean | 101.73068690299988 | 99.33992314338684 | -2.37803% |
| Null | 0.9008660316467285 | 0.9387650489807129 | |
| | 0.9166581630706787 | 0.8799121379852295 | |
| | 0.905055046081543 | 0.8913090229034424 | |
| | 0.9088449478149414 | 0.9316670894622803 | |
| | 0.9060719013214111 | 0.9277148246765137 | |	
| Avg. | 0.90749921798706 | 0.91387362480164 | +0.699956% |

#### And some benches on building Dune:

##### Main
```
Benchmark 1: ./dune.exe build @install --cache disabled
  Time (mean ± σ):      9.826 s ±  0.280 s    [User: 90.914 s, System: 28.879 s]
  Range (min … max):    9.290 s … 10.279 s    10 runs
```
##### PR
```
Benchmark 1: ./dune.exe build @install --cache disabled
  Time (mean ± σ):     10.206 s ±  0.342 s    [User: 91.557 s, System: 29.759 s]
  Range (min … max):    9.916 s … 11.012 s    10 runs
```
Both results tell me that there still does appear to be an observable slow down.