@rgrinberg 2023-05-27
@snowleopard this option is now under config.